remove outsourced PRs
