+++ backend/lib/piece-indexer.js
@@ -1,5 +1,9 @@
-import { storeActiveDeals } from '@filecoin-station/deal-observer-db/lib/database-access.js'
 import { fetchPayloadCid } from './pix-service/service.js'
+import { loadDeals, storeActiveDeals } from './deal-observer.js'
+
+/** @import {Queryable} from '@filecoin-station/deal-observer-db' */
+/** @import { Static } from '@sinclair/typebox' */
+/** @import { ActiveDealDbEntry } from '@filecoin-station/deal-observer-db/lib/types.js' */
 
 /**
  * @param {import("@filecoin-station/deal-observer-db").Queryable} pgPool
@@ -17,4 +21,23 @@ export async function updatePayloadCids (pgPool, makeRpcRequest, activeDeals, pi
   if (updatedDeals.length > 0) {
     await storeActiveDeals(updatedDeals, pgPool)
   }
+}
+
+export const pieceIndexerLoopFunction = async (rpcRequest, pixRequest, pgPool, queryLimit) => {
+  // TODO: handle payloads which cannot be retrieved from the piece CID indexer
+  const dealsWithMissingPayloadCid = await fetchDealsWithNoPayloadCid(pgPool, queryLimit)
+  if (dealsWithMissingPayloadCid !== null && dealsWithMissingPayloadCid) {
+    await updatePayloadCids(pgPool, rpcRequest, dealsWithMissingPayloadCid, pixRequest)
+  }
+}
+
+/**
+   * @param {Queryable} pgPool
+   * @param {number} limit
+   * @returns {Promise<Array<Static< typeof ActiveDealDbEntry>>>}
+   */
+export async function fetchDealsWithNoPayloadCid (pgPool, limit) {
+  const query = `SELECT * FROM active_deals WHERE payload_cid IS NULL ORDER BY activated_at_epoch ASC LIMIT ${limit}`
+  const result = await loadDeals(pgPool, query)
+  return result
 }
