+++ backend/lib/pix-service/service.js
@@ -2,13 +2,7 @@ import { Value } from '@sinclair/typebox/value'
 import { PIECE_INDEXER_URL } from '../config.js'
 import { getMinerPeerId } from '../rpc-service/service.js'
 import { PixResponse } from './data-types.js'
-import { cacheOptions } from './utils.js'
 import pRetry from 'p-retry'
-import { LRUCache } from 'lru-cache'
-
-// A cache for miner peer IDs
-// It has a maximum of 10000 items and deletes the least recently used items
-const minerPeerIdsCache = new LRUCache(cacheOptions)
 
 /**
  * @param {string} providerId
@@ -30,15 +24,7 @@ export const pixRequest = async (providerId, pieceCid) => {
 * @returns {Promise<string>}
 */
 export async function fetchPayloadCid (providerId, pieceCid, rpcRequest, pixRequest) {
-  let minerPeerId
-  // Check for cached miner peer IDs
-  if (minerPeerIdsCache.has(providerId)) {
-    minerPeerId = minerPeerIdsCache.get(providerId)
-  } else {
-    // If the miner peer ID is not cached, fetch it from the rpc service
-    minerPeerId = await getMinerPeerId(providerId, rpcRequest)
-    minerPeerIdsCache.set(providerId, minerPeerId)
-  }
+  const minerPeerId = await getMinerPeerId(providerId, rpcRequest)
   const payloadCid = await pixRequest(minerPeerId, pieceCid)
   return payloadCid
 }
