+++ backend/package.json
@@ -19,7 +19,6 @@
     "@sentry/node": "^8.48.0",
     "@sinclair/typebox": "^0.34.14",
     "debug": "^4.4.0",
-    "lru-cache": "^11.0.2",
     "multiformats": "^13.3.1",
     "p-retry": "^6.2.1",
     "pg": "^8.13.1",
