+++ backend/test/rpc-client.test.js
@@ -1,13 +1,24 @@
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
 import { chainHeadTestData } from './test_data/chainHead.js'
+import { rawActorEventTestData } from './test_data/rawActorEvent.js'
 import { parse } from '@ipld/dag-json'
-import { getActorEvents, getActorEventsFilter, getChainHead, getMinderInfoParameters, rpcRequest } from '../lib/rpc-service/service.js'
+import { getActorEvents, getActorEventsFilter, getChainHead, rpcRequest } from '../lib/rpc-service/service.js'
 import { ClaimEvent } from '../lib/rpc-service/data-types.js'
 import { Value } from '@sinclair/typebox/value'
-import { makepixRequest, makeRpcRequest } from './utils.js'
 
 describe('RpcApiClient', () => {
+  const makeRpcRequest = async (method, params) => {
+    switch (method) {
+      case 'Filecoin.ChainHead':
+        return parse(JSON.stringify(chainHeadTestData))
+      case 'Filecoin.GetActorEventsRaw':
+        return parse(JSON.stringify(rawActorEventTestData)).filter(e => e.height >= params[0].fromHeight && e.height <= params[0].toHeight)
+      default:
+        console.error('Unknown method')
+    }
+  }
+
   it('retrieves the chainHead', async () => {
     const chainHead = await getChainHead(makeRpcRequest)
     assert(chainHead)
@@ -34,21 +45,4 @@ describe('RpcApiClient', () => {
     const actorEvents = await getActorEvents(getActorEventsFilter(chainHead.Height - 1000, 'claim'), rpcRequest)
     assert(actorEvents)
   })
-
-  for (let blockHeight = 4622129; blockHeight < 4622129 + 11; blockHeight++) {
-    it('retrieves minderInfos and payloadCids in block {blockHeight}', async () => {
-      const actorEvents = await getActorEvents(getActorEventsFilter(blockHeight, 'claim'), makeRpcRequest)
-      for (let i = 0; i < actorEvents.length; i++) {
-        const e = actorEvents[i]
-        const parsedEvent = Value.Parse(ClaimEvent, e.event)
-        const params = getMinderInfoParameters(parsedEvent.provider)
-        const res = await makeRpcRequest('Filecoin.StateMinerInfo', params)
-
-        assert(res)
-        assert(res.PeerId)
-        const payloadCid = await makepixRequest(parsedEvent.provider, parsedEvent.pieceCid)
-        assert(payloadCid)
-      }
-    })
-  }
 })
